Map the ROM address into the Memory of the NES
This will ensure that the header of the ROM is ignored, and read/write
operations land in the correct address space of the ROM.

The header is of size 0x0010, and the code that follow is mapped to
address 0x8000. For example, reading 0x80A0 will read the data in the
ROM at offset 0x00B0.

Test: bazel run src:atlas_test