Move the instructions to a separate directory
This begins refactoring the instructions away from the Cpu, so that they
can be developed and tested independently. There is still some cleanup
to do, such as removing the Cpu dependency from Instruction.

One way of doing this would be to introduce a CpuProxy interface that
Instructions use to get/set the memory/registers. The Cpu would
initialize a custom version of this interface and inject it into each
Instruction upon creation.