Remove Instruction's dependency on Cpu/Memory
A new interface CpuProxy is introduced to clean up the dependencies, so
that Instruction no longer depends on Cpu and Memory, which are more
concrete classes.

Instructions use CpuProxy to get/set the memory and registers without
needing the entire cpu.h/memory.h definitions. Cpu implements a CpuProxy
and injects it into each Instruction upon construction.

A future consideration is whether Cpu should implement CpuProxy
directly, rather than instantiating a new class. Also, potentially
cpu.h and memory.h could be slimmed down to the bare minimum, so that it
is actually appropriate for Instructions to depend on them.