Abstract Memory behind a Peripheral class
We want to add a class to manage the PPU, which receives Read/Write
commands for a select address range. This logic does not belong in the
Memory class, so we are introducing a new Bus/Peripheral abstraction, so
that the PPU can sit beside the Memory code.

Most classes will not invoke the PPU/Memory directly anymore, but will
instead go through the Bus.

This commit starts the migration by putting Memory behind the Peripheral
interface. Next step is to create the Bus class to handle the
redirection of all Read/Write operations to the appropriate Peripheral.