Prepare glad/glfw/glm projects
This commit downloads and builds the glad/glfw/glm projects for helping
with graphics support. For each of these projects, cmake is used to
build.

cmake is a new dependency, and potentially python for glad. I am not
fully sure of all the hidden dependencies with this change. I'd have to
test on a clean VM to know for sure.