Render the tiles with palette colors using opengl
This massive commit pulls in the opengl code to start rendering the
nametable. It uses the correct color palettes, so that we actually have
an initial frame that looks like a game :D

There is a lot of opportunity for optimization after this commit. For
example, this commit currently does these nasties:

1) Loads 960 textures (1 for every tile in the view), and does not
support scaling to multiple namespaces.

2) Combines the pattern + color palette for every texture 60
times/second, no matter if they never changed.

3) Reloads every texture 60 times/second (the refresh rate), no matter
if they never changed.