Migrate clock to use std::chrono
I found that my math was incorrect in the clock code, so I opted to use
the std::chrono library to catch errors in unit casting during compile
time. This also makes the code look a bit cleaner, and more portable
across platforms.

This commit introduces a problem...we read the current time using
std::chrono::steady_clock::now(), which is a syscall, instead of
tracking our ticks manually in a uint64_t. While this seems all find and
dandy, syscalls are expensive, and this causes the entire emulator to
run at a snails pace.

Before pushing this to the main branch, I would like to fix this.