Replace Clock with faster hardcode in Atlas
The Clock was slowing everything down by doing a syscall (fetching
current time) for every CPU cycle. In order to fix this, we are running
an entire frames-worth of CPU cycles in a row, rendering with the PPU,
then sleeping once. This essentially leads to 2 syscalls per-frame,
which is significantly faster.