Only load modified PPU data
During a frame render, we now only reload the tables that have modified.
For example, if only the frame palette has been modified since the last
render, then only the frame palette will be updated in the window.

This has the effect of speeding up the NOP benchmark from 1100us to 51us.