Display the performance in the window title
This includes the clock speed in terms of percentage of expected speed.
We desire this to be 100%. Lower than 100% means we are running slow.
Also, we include the idle time, which is the percentage of time spent
sleeping. We desire this to be as low as possible. And idle percentage
of 0% means we are not sleeping at all.