Add vertical/horizontal flipping for sprites
Currently this is done by swapping the texture coordinates around. Maybe
in the future we could consider adding the ability to modify sprite
attributes without re-computing the entire tile, vertices, elements, and
texture coordinates each time. This would require us to have additional
logic in the PPU to detect when only an attribute has changed, and not
the tile number.