Scroll by modifying texture coordinates
Instead of using a uniform in the vertex shader to scroll through
several quads, we now have a single quad (ignoring the zero-color quad),
that uses texture coordinates between 0.0 .. 3.0 to select the correct
pixels. When scolling, the texture coordinates are modified. The
fragment shader converts the 0.0 .. 3.0 coordinates to the correct table
number.

Currently, this only supports horizontal mirroring, but this is the
first step for split rendering.