Detect sprite 0 hit mid-frame
One each scanline, we check if sprite 0 has been hit, so that the CPU
can do raster timing mid-frame. This is required in order to facility
partial horizontal scrolling.

While testing this, I found that some ROMs depend on physical behavior
in the PPU that causes writing to PpuAddr to affect the scroll and
base nametable. This commit also adds the logic to accomodate for that.