Detect a vertical split and inform the Renderer
Vertical splits can be detected by monitoring the PpuAddr and PpuScroll,
while the PPU is running through the visible scanlines. If the
PpuScroll.x is modified during the visible scanlines, then we have a
vertical split with a partial horizontal scroll. If the PpuAddr is
modified during the visible scanlines, then we have a vertical split
with a partial vertical scroll.