Combine all addressing modes into single class
The amount of code did not warrant separate classes, and this cleans up
the logic, and makes it more readable.

Addressing is now owned by the Cpu.

The specific mode can be selected with a new enum, and the Engine uses
this enum during instruction registration.

We can no longer have a |nullptr| mode, which is good for memory safety.