Add Renderer2 with an implementation and test
In order to support more renderer implementations, it would be ideal to
make the interface as simple as possible. Also, looking forward to other
emulators, such as GBA, there's some tricky rendering hacks that ROMs
use, which we do not want to re-implement for every renderer
implementation. An example of this is ROMs changing the frame layout
type mid-frame.

We are moving all of the business logic out of the Renderer and into a
new PPU implementation. The Renderer will now be responsible for drawing
a bag of bytes and that's it.

This CL introduces a Renderer2 interface, OpenGL implementation, and a
test to verify it is working.